Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { View, FlatList, Text } from "react-native";
+import { View, FlatList, Text, Alert } from "react-native";
 import { ListItem } from "react-native-elements";
 import { connect } from "react-redux";
 import { baseUrl } from "../shared/baseUrl";
@@ -30,7 +30,27 @@ class Favorites extends Component {
         {
           text: "Delete",
           type: "Delete",
-          onPress: () => this.props.deleteFavorite(item.id),
+          onPress: () => {
+            Alert.alert(
+              "Delete Favorite?",
+              "Are you sure you want to delete this favorite dish " +
+                item.name +
+                "?",
+              [
+                {
+                  text: "Cancel",
+                  onPress: () => console.log(item.name + " Not Deleted"),
+                  style: "cancel",
+                },
+                {
+                  text: "Yes",
+                  onPress: () => this.props.deleteFavorite(item.id),
+                  style: "default",
+                },
+              ],
+              { cancelable: false }
+            );
+          },
         },
       ];
